Disable __del__ functions
Due to errors that occur when terminating scripts that make use of these classes.
Apparently, after the last statement of a script is executed, the python process starts destryoing objects in certain order (that I don't fully understand yet) but it causes the self.cnx object to break down before the self.cnx.close() statement is executed.